[enh] Move view tool services to service architecture
