+++ backend/views/syson-diagram-common-view/src/main/java/org/eclipse/syson/diagram/common/view/edges/AnnotationEdgeDescriptionProvider.java
@@ -26,7 +26,7 @@ import org.eclipse.sirius.components.view.diagram.EdgeStyle;
 import org.eclipse.sirius.components.view.diagram.LineStyle;
 import org.eclipse.sirius.components.view.diagram.NodeDescription;
 import org.eclipse.sirius.components.view.diagram.SynchronizationPolicy;
-import org.eclipse.syson.diagram.common.view.services.ViewToolService;
+import org.eclipse.syson.diagram.services.aql.DiagramMutationAQLService;
 import org.eclipse.syson.services.UtilService;
 import org.eclipse.syson.sysml.SysmlPackage;
 import org.eclipse.syson.util.AQLConstants;
@@ -108,7 +108,7 @@ public class AnnotationEdgeDescriptionProvider extends AbstractEdgeDescriptionPr
     @Override
     protected ChangeContextBuilder getTargetReconnectToolBody() {
         return this.viewBuilderHelper.newChangeContext()
-                .expression(ServiceMethod.of1(ViewToolService::reconnnectTargetAnnotatedEdge)
+                .expression(ServiceMethod.of1(DiagramMutationAQLService::reconnnectTargetAnnotatedEdge)
                         .aql(AQLConstants.EDGE_SEMANTIC_ELEMENT, AQLConstants.SEMANTIC_RECONNECTION_TARGET));
     }
