+++ backend/views/syson-standard-diagrams-view/pom.xml
@@ -101,6 +101,10 @@
 		</dependency>
 		<dependency>
 			<groupId>org.eclipse.syson</groupId>
+			<artifactId>syson-tree-services</artifactId>
+		</dependency>
+		<dependency>
+			<groupId>org.eclipse.syson</groupId>
 			<artifactId>syson-common-view</artifactId>
 		</dependency>
 		<dependency>
